+++ pakyow-core/lib/pakyow/application/behavior/rescuing.rb
@@ -5,33 +5,38 @@ require "pakyow/support/extension"
 module Pakyow
   class Application
     module Behavior
-      # Lets an app to be rescued from errors encountered during boot. Once rescued,
-      # an app returns a 500 response with the error that caused it to fail.
+      # Lets an application to be rescued from errors encountered when running. Once rescued,
+      # the application triggers a 500 response with the error that caused it to fail.
       #
       module Rescuing
         extend Support::Extension
 
-        common_methods do
-          # Error the app was rescued from.
-          #
-          def rescued
-            defined?(@rescued) ? @rescued : self.class.rescued
-          end
+        apply_extension do
+          events :rescue
+        end
 
-          # Returns true if the app has been rescued.
+        common_methods do
+          # Error the application was rescued from.
+          #
+          def error
+            defined?(@error) ? @error : self.class.error
+          end
+          alias rescued error
+
+          # Returns true if the application has been rescued.
           #
           def rescued?
             if is_a?(Application)
-              self.class.rescued? || (instance_variable_defined?(:@rescued) && !!@rescued)
+              self.class.rescued? || (instance_variable_defined?(:@error) && !!@error)
             else
-              instance_variable_defined?(:@rescued) && !!@rescued
+              instance_variable_defined?(:@error) && !!@error
             end
           end
 
           # Enters rescue mode after reporting the error.
           #
           def rescue!(error)
-            @rescued = error
+            @error = error
 
             performing :rescue, error: error do
               Pakyow.houston(error)
@@ -47,10 +52,9 @@ module Pakyow
 
         module Rescued
           def call(connection)
-            isolated(:Connection).new(self, connection).tap do |application_connection|
-              application_connection.error = rescued
-              application_connection.trigger 500
-            end
+            application_connection = isolated(:Connection).new(self, connection)
+            application_connection.error = rescued
+            application_connection.trigger 500
           end
         end
       end